refactor: delete candiate password change and company password change
