+++ src/main/java/com/ctrls/auto_enter_view/controller/CompanyController.java
@@ -1,6 +1,5 @@
 package com.ctrls.auto_enter_view.controller;
 
-import com.ctrls.auto_enter_view.dto.company.ChangePasswordDto;
 import com.ctrls.auto_enter_view.dto.company.SignUpDto;
 import com.ctrls.auto_enter_view.dto.company.WithdrawDto;
 import com.ctrls.auto_enter_view.enums.ResponseMessage;
@@ -11,7 +10,6 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,17 +27,6 @@ public class CompanyController {
     SignUpDto.Response response = companyService.signUp(form);
 
     return ResponseEntity.ok(response);
-  }
-
-  // 비밀번호 수정
-  @PutMapping("/companies/{companyKey}/password")
-  public ResponseEntity<String> changePassword(
-      @PathVariable String companyKey,
-      @Validated @RequestBody ChangePasswordDto.Request form) {
-
-    companyService.changePassword(companyKey, form);
-
-    return ResponseEntity.ok(ResponseMessage.CHANGE_PASSWORD.getMessage());
   }
 
   // 회원 탈퇴
