--- src/main/java/com/ctrls/auto_enter_view/dto/company/ChangePasswordDto.java
@@ -1,23 +1,0 @@
-package com.ctrls.auto_enter_view.dto.company;
-
-import jakarta.validation.constraints.Pattern;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class ChangePasswordDto {
-
-  @Getter
-  @Builder
-  @NoArgsConstructor
-  @AllArgsConstructor
-  public static class Request {
-
-    @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}", message = "비밀번호 형식이 올바르지 않습니다.")
-    private String oldPassword;
-
-    @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}", message = "비밀번호 형식이 올바르지 않습니다.")
-    private String newPassword;
-  }
-}
