+++ src/main/java/com/ctrls/auto_enter_view/service/CandidateService.java
@@ -6,7 +6,6 @@ import static com.ctrls.auto_enter_view.enums.ErrorCode.PASSWORD_NOT_MATCH;
 import static com.ctrls.auto_enter_view.enums.ErrorCode.USER_NOT_FOUND;
 
 import com.ctrls.auto_enter_view.dto.candidate.CandidateApplyDto;
-import com.ctrls.auto_enter_view.dto.candidate.ChangePasswordDto.Request;
 import com.ctrls.auto_enter_view.dto.candidate.FindEmailDto;
 import com.ctrls.auto_enter_view.dto.candidate.FindEmailDto.Response;
 import com.ctrls.auto_enter_view.dto.candidate.SignUpDto;
@@ -85,28 +84,6 @@ public class CandidateService {
     }
 
     candidateRepository.delete(candidateEntity);
-  }
-
-  // 비밀번호 수정
-  public void changePassword(String candidateKey, Request request) {
-
-    User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-    CandidateEntity candidateEntity = candidateRepository.findByEmail(principal.getUsername())
-        .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
-
-    // 응시자 정보의 응시자키와 URL 응시자키 일치 확인
-    if (!candidateEntity.getCandidateKey().equals(candidateKey)) {
-      throw new CustomException(USER_NOT_FOUND);
-    }
-
-    if (!passwordEncoder.matches(request.getOldPassword(), candidateEntity.getPassword())) {
-      throw new CustomException(PASSWORD_NOT_MATCH);
-    }
-
-    candidateEntity.setPassword(passwordEncoder.encode(request.getNewPassword()));
-
-    candidateRepository.save(candidateEntity);
   }
 
   public Response findEmail(FindEmailDto.Request request) {
